Move site options into new lib folder
