+++ apps/full-site-editing/full-site-editing-plugin/full-site-editing/blocks/site-description/edit.js
@@ -29,7 +29,7 @@ import { PanelBody } from '@wordpress/components';
 /**
  * Internal dependencies
  */
-import useSiteOptions from '../useSiteOptions';
+import { useSiteOptions } from '../../lib';
 
 function SiteDescriptionEdit( {
 	attributes,
