+++ apps/full-site-editing/full-site-editing-plugin/full-site-editing/blocks/site-title/edit.js
@@ -28,7 +28,7 @@ import { PanelBody } from '@wordpress/components';
 /**
  * Internal dependencies
  */
-import useSiteOptions from '../useSiteOptions';
+import { useSiteOptions } from '../../lib';
 
 function SiteTitleEdit( {
 	attributes,
