+++ apps/full-site-editing/full-site-editing-plugin/full-site-editing/lib/index.js
@@ -1,0 +1,1 @@
+export * from './site-options';
