+++ apps/full-site-editing/full-site-editing-plugin/full-site-editing/lib/site-options/index.js
@@ -1,0 +1,1 @@
+export * from './use-site-options';
